1.2.0: Improve interface and memory management
Version 1.2.0 gets rid of the use of arrays as the data-holding variables
and replaces them with vectors.  Also, the user interface is slightly improved
for usability:  More verbose instructions, ability to keep making new images
without the application exiting each time, and the coloring of the set is more
logical (I.E. no more normalizing color schemes based on the iteration number).
Also the generated image is now a fixed size and the user can only change the
center coordinates and the zoom level.