Standardize syntax and fix broken strings
After the last code cleanup commit there were some errors involving the strings
that are displayed during execution.  Mainly the way the strings were broken up
over multiple lines was throwing errors, so this commit goes back to the
original wall of text that was present in 1.2.0 to fix that.

Also the indenting is consistent now throughout all files:  Tabs are for
separating structures and spaces are used to align statements within those
structures.  This helps make the code easier to read across systems that use
different sized tab indents.

Also the comment format is consistent throughout the files now.